+++ core/src/se/tevej/game/model/resource/Resource.java
@@ -1,15 +1,15 @@
 package se.tevej.game.model.resource;
 
 public class Resource {
-    private float amount;
+    private double amount;
     private ResourceType type;
 
-    public Resource(float amount, ResourceType type) {
+    public Resource(double amount, ResourceType type) {
         this.amount = amount;
         this.type = type;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
